pyproject.toml
[project]
name = "webpage-guard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.90.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
]
